remove unnecessary float cast
git-svn-id: file:///home/svn/shoola-omero/trunk@1723 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/chainbuilder/piccolo/ChainView.java
@@ -437,7 +437,7 @@ public class ChainView extends PNode implements BufferedObject, MouseableNode {
 	
 	public double getHeight() { 
 		//return chainHeight;
-		return (float) chainBounds.getHeight()+VGAP;
+		return  chainBounds.getHeight()+VGAP;
 	}
 	
 	public void adjustVerticalExtents(PBounds b) {
